AI/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.1.0
branca==0.8.1
Brotli==1.0.9
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datascience==0.17.6
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.8
filelock==3.13.1
flatbuffers==25.2.10
folium==0.19.4
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
kobert-tokenizer==0.1
langchain==0.3.18
langchain-core==0.3.35
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.8
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.25.2
nest_asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.63.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==3.19.6
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
stack_data==0.6.3
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.6
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
